requirements.txt
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
openai
python-dotenv
beautifulsoup4
requests